requirements.txt
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
debugpy @ file:///croot/debugpy_1736267418885/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1733327148154/work
et_xmlfile==2.0.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
google-auth==2.40.3
google-auth-oauthlib==1.2.2
greenlet==3.2.3
gspread==6.2.1
httplib2==0.22.0
idna==3.10
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1748711175/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1748333051527/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mysql-connector-python==9.3.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1653089181607/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyodbc==5.2.0
pyparsing==3.2.3
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.1.0
pytz==2025.2
pyzmq @ file:///croot/pyzmq_1734687138743/work
RapidFuzz==3.13.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
SQLAlchemy==2.0.41
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827254365/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1748959427/work
tzdata==2025.2
urllib3==2.4.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
